backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis[asyncio]==5.0.4
celery==5.3.6
anthropic==0.25.0
voyageai==0.2.3
neo4j==5.19.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
supabase==2.4.5
markdown==3.6
python-dotenv==1.0.1
bcrypt==4.2.1
slowapi==0.1.9
structlog==24.4.0
sentry-sdk[fastapi]==2.14.0
resend==2.0.0
